Added support to open files from a url
Added a function to the Convertor class that will check and see if the
INPUT_FILE argument is a url. If the argument is a valid url, then the
Python urllib library will be used to request the file and open it and
covnert the file to an org-mode file without the need to download the file.